CI: add refl-cpp to singleheader dependencies
The amalgamate.py script needs the include directories it should merge
into the single header specified in it's
`devtools/singleheader-config-*.json` configuration file.

This commit adds the refl-cpp include directory to the graph single
header. The dependeny was there before, but it is only used now, which
is why this surfaced.