Use do-not-optimize inline-asm for more realistic results
1. If an inline-asm statement clobbers all memory then we don't see
realistic benchmark measurements because the optimizer is inhibited from
doing certain optimizations. The result is very different depending on
the order the node's process_one functions are called. The constexpr
cases suffered most because the optimizer was inhibited in every single
iteration.

2. The source produces only zero-initialized samples. The optimizer was
able to see that. And it may or may not have used that knowledge,
rendering all previous results as quasi-random. Use inline asm to only
inhibit const-prop, nothing else.

ChangeLog:

	* bench/benchmark.hpp (do_not_optimize): Optimal do-not-optimize
	implementation. Do not invalidate all memory in the process.
	* bench/bm_test_helper.hpp (source::process_one): Taint the
	constructed sample using benchmark::do_not_optimize.