Don't use inputs_status for source nodes
node::inputs_status reports whether, and how much data is available on
the input ports of a node. However, source nodes have no inputs and
therefore the call makes no sense (it was reporting 1 sample before,
which is nonsense).

Now, node::work will only call inputs_status if the node is not a
source. Otherwise, it will first try to find a customization point
(available_samples). If that doesn't exist and the node is also a sink,
it looks through all the internal ports that were merged, whether any
specifies a maximum sample size. If yes it uses the minimum value. If
it's not a sink, then it uses the available output buffer size.

The benchmark source node implements the customization point
available_samples. It reports as many samples available as were
requested and not produced yet.

ChangeLog:

	* bench/bm_test_helper.hpp (source::available_samples): New.
	* include/node.hpp (node::enough_samples_for_output_ports): New.
	(node::space_available_on_output_ports): New.
	(node::inputs_status): Ensure the function is not called on
	source nodes. Remove one level of lambda.
	(node::work): If this node is a source node determine the number
	of samples to process via available_samples, max_samples from
	internal ports, or available output buffer size. Otherwise,
	query inputs_status, as before.
	(merged_node::merged_work_chunk_size): New.
	(merged_node::available_samples): New.
	* include/port.hpp (port::reader): Fix const on return type.
	(port::writer): Likewise.
	* include/port_traits.hpp (traits::port::min_samples): New.
	(traits::port::max_samples): New.