Clean up cache line alignment usage
1. Use std::hardware_destructive_interference_size when available,
otherwise hardcode the value to 64. This warns about making the ABI
depend on -mtune/-mcpu, which is true. Anyone have a problem with this?

2. Use over-alignment sparingly. The cost is in terms of bandwidth, TLB
and cache. However, values that are accessed together should be in the
same cache line. Therefore, align complete objects such that their data
members are never split over a cache line boundary.

ChangeLog:

	* include/circular_buffer.hpp (buffer_impl): Remove alignas.
	The Sequence data member is over-aligned, therefore put it
	first.
	(buffer_writer): Remove alignas.
	(buffer_reader): Remove alignas.
	* include/claim_strategy.hpp (SingleThreadedStrategy):
	Over-align the complete type but not its members.
	(MultiThreadedStrategy): Likewise.
	* include/sequence.hpp
	(hardware_destructive_interference_size): Replaces kCacheLine.
	(hardware_constructive_interference_size): Add.
	* test/qa_buffer.cpp: Likewise.