basic node setting handling
Settings are public node member fields (<-> PoCo) that can be accessed directly (default from C++ point-of-view) or via key-pmt-value-map that can be interacted via 'setting().set(..)' and 'setting().get(..)'.

The settings are thread-safe and real-time/non-real-time compatible and changed either directly during the node initialisation or in the 'work()' function prior to calling the user's 'process_[one, bulk](...)' functions.

The setting management requires that the fields are registered using the compile-time reflection extension macro.

This implementation is pluggable (i.e. user can provide their own setting handling logic) and has API provisions for optional transaction and timing/ctx-based settings management that is to be completed with a later PR.

Signed-off-by: rstein <r.steinhagen@gsi.de>