node::work(): add missing case for sinks
If available_samples returns 0, the node has to immediately return with
OK, because otherwise it tries to get a writer for 0 entries, triggering
an assertion in `claim_strategy::next`.
The `space_available_on_output_port()` check never fired because the
code above limits `samples_to_process` to the available numbers, so if
samples to process is zero and available_samples nonzero, the samples
have been limited to zero, indicating a full output buffer.