Transfer and improve opencmw::TimingCtx
This is by large a port of the opencmw TimingCtx, but with some
improvements.

The most significant change is that the parse and match functions are
not hardcoded anymore, but can instead be passed over from the outside
via a predicate.
This is necessary to ensure that TimingCtx stays generic enough so that
it can be used for arbitrary usecases.

Furthermore, we no longer hardcode any ids but instead use a
`pmtv::map_t` to allow the user to store arbitrary identifiers for their
usecase.