Transfer and improve opencmw Transaction SettingBase
This required some extensive changes and adaptions to work with the more
general version of TimingCtx and to use the circular buffer instead of
the opencmw::RingBuffer.

This also adds a new Utils function to find out the true cache line
size. It uses `std::hardware_destructive_interference_size` and falls
back to 64, which is the cache line size for most x86_64 targets.

Now the very last piece in the puzzle is to port the CtxSetting class.