Work around Clang bug
Clang either fails to evaluate the requires expression correctly or it
fails to not instantiate the constexpr-if else branch. This lead to
ill-formed calls which shouldn't be possible.

The workaround splits the tuple_for_each function into two different
functions, explicitly requesting whether the index parameter should be
added or not.

Use the new tuple_for_each_enumerate function in Node::write_to_outputs.