Refactor Transactions
The old implementation used redundant locking mechanisms and
unnecessarily complicated circular buffers.

Instead this now uses just a single recursive lock and a normal
collection, which simplifies this a lot and hopefully makes emscripten
more happy as well.

This also integrates the multiplexed settings into the Node as a drop-in
replacement for the basic settings.