Fix reallocating Port instances
The move constructor was accidentally missing some fields, which caused
all connections to break on reallocation.
This is the reason why the qa_DynamicBlock test could not use an
std::vector to hold the input ports and instead worked around the
problem by using a std::list.

With the source of the problem fixed now, we switch the test back to
using a std::vector.

Note that this is not the final end of the story. The relationship
between DynamicPort and Port is still a bit of a complicated one,
because the lifetime of these is so tightly coupled together.
DynamicPort objects store a reference to their Port counterpart, so
reallocating a Port without informing (or recreating) a DynamicPort will
quickly lead to problems.

A possible way to fix this in the future is to add a callback to Port,
that will be called on its move constructor to allow the DynamicPort to
adapt its reference to Port.

Fixes https://github.com/fair-acc/opendigitizer/issues/90