multi_adder: Consume inputs outside of an assert()
This one was the kind of bug that makes any programmer have nightmares:
The test failed only on release builds, it worked fine on Debug builds,
where you would want to actually debug the problem. Looks like
printf-style debugging is back on the menu.

Turns out that an essential line was wrapped in an assert() which of
course is completely disabled in release builds.