connect with index is not always deprecated
We use index-based connect internally, so whichever API the user
uses, it will end up in the index-based connect and report a warning.

Additionally, for the default message ports, the only way to
connect them is index-based.

This patch makes the deprecated warnings only shown if the user
tries to use the index-based connect directly with the index not
being default_message_port_index.