Create the maximum number of threads when the task type is CPU_BOUND
This fixes the scheduler not working properly because it divides
the blocks among the maximum number of threads but the thread pool
was only creating the minimum amount of threads.
When minimum and maximum are not the same that resulted in only part
of the blocks being executed, in case of a flowgraph that keeps running.