CircularBuffer: fix MultipleProducer claim strategy
The buffer writer was is updating the available flag buffer but the
reader logic of CircularBuffer would read until cursor, which
corresponds to the position of the next available slot and also includes
slots that have not yet been published, leading to race conditions.

This gets the available samples to read by calling the
getHighestPublishedSequence method in reader::available.
Also the publish function was fixed to make all published entries
available instead of just the last one.

Also corrects an off by one error in the required bitshift for the
available flag bitshift.

Useas an alternative implementation on libc++ based builds as atomic_ref
is not supported by libc++.