Do not register block at init time
Revert the renaming of Block::init() to Block::performInit().

Instead of registering the DataSink at init time, do so in start(),
and unregister in stop().

This needs changes in the tests, as a poller/callback can now only
be registered while the execution is already running, the
pollers/callbacks might not get all the data from the start.
Add a initial delay (new block testing::Delay) to work around that.