CI: enable uploading coredumps and coresponding binaries
This adds CI steps to enable the collection of coredump files and
uploading them as an artefact together with the binary that caused them
as an artefact for offline analysis.
The general steps where taken from
https://github.com/itamarst/gha-upload-cores
but adapted to our more complex repository setup with multiple
executables in different directories.