Block: fix EoS handling for process bulk not consuming samples
If the processBulk function does not consume all samples up to the eos
tag, the previous code would terminate processing anyway, leading to the
unprocessed samples being dropped.
This is becaue the ConsumableSpan does not provide a way to determine
how much the processsing function has actually consumed.

This change solves this by only handling explicitly stopped blocks after
the block has processed, which leaves the eos tag handling for the start
of the next call, where it is sufficient to check if the eos tag is the
first sample.

Also fixes a few minor inconsistencies discovered by @frankosterfeld.