Core: rework Tag handling
The automatic tag forwarding implemented in Block::workInternal was was not working for
ASYNC ports since in this case the Block cannot determine a priori if samples will
be consumed and consequently cannot perform an early return ahead of the unconditional
tag forwarding and consumption.

This change introduces an extension of the ConsumableInputRange, which additionally takes
care of consuming all tags up to the first sample of the currently processed chunk.

Also all tag merging logic is moved from Port to Block, while the logic for preparing the
consumableInput or producableOutput ranges is moved into the Port.

Only forward initial tags before starting the flowgraph instead of
during initialisation, when the buffers are not connected yet.

In CircularBuffer, make PublishableOutputRange pubblish in destructor

This makes the PublishableOutputRange behave like the corresponding
InputRange by only marking samples for being published and only
publishing in the destructor.

Evaluate actually processed samples after process bulk was called.

Add port unittest and fix off by one problem

Change sequence from -1 to 0 to avoid tag offset

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>