Fix optional gcc dependency on TBB
For implementing std::execution::par, gcc relies on the TBB library.
On systems without TBB installed it falls back to the non-parallel
implementation, but as soon as the TBB header is available it will try
to use it and require linking against TBB.
https://github.com/gcc-mirror/gcc/blob/releases/gcc-14.1.0/libstdc%2B%2B-v3/include/bits/c%2B%2Bconfig#L874-L882
Since TBB is a bigger dependency, making it a required dependency is not
an option, so this introduces the option ENABLE_TBB which by default is
off but can be used to add the dependency.

fixes #361

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>