Notes from merge-API discussion with Ralph
Settings are now propagated to the 'left' and 'right' members of
MergedGraph.

Add TODO comments about:

1. sizeof(MergedGraph<A, B, ...>) is larger than necessary. Most Block
base and direct data members of A and B are not used. They are only
needed for the MergedGraph specialization itself (which derives from
Block). This came up when considering the state of a block. Each member
block of a MergedGraph has incorrect state information. Instead the
should be only one state for the complete block and all sub-blocks
should see that state.

2. Not every Block type should be mergeable. Heuristic: if processOne is
'const', then yes. Otherwise an explicit opt-in seems to be our best
option.