CMake: Default color diagnostics to force-on with ccache/ninja
Reading diagnostics with ANSI color codes in between is unnecessarily
hard. We now make use of the cmake documented switch to control color
output. And we only force color on if the tooling suggest it might be
necessary.

Since GitHub Actions run without a terminal, auto-detection would always
determine to not use colored output. However, color works fine on GitHub
and therefore we force it on in CI builds.

Signed-off-by: Matthias Kretz <m.kretz@gsi.de>