algorithm: fix gcc warnings in fftw wrapper
gcc is giving -Wsubobject-linkage warnings when gr::algorithm::FFTw is
used as a member of a block because it says that fields such as
fftwIn have internal linkage. This is caused by a lambda in the deleter
of the std::unique_ptr used in InUniquePtr, etc. The warning can be
removed by using a deleter struct instead of a lambda.

Signed-off-by: Daniel Estévez <daniel@destevez.net>