CircularBuffer: do not reset empty PublishableOutputRange
When an empty PublishableOutputRange is requested, do not reset
the index, offet and internal span of the buffer.

This solution was suggested by @drslebedev to fix bugs in which
the buffer state would become corrupted and report negative values
(casted to uint64_t) for the available number of elements, etc.

Signed-off-by: Daniel Estévez <daniel@destevez.net>