Remove non-working call to blocks->ports->publishPendingTags
The call was calling the method on the scheduler itself instead of on
the blocks it was iterating over. This was probably an incomplete
refactoring in moving this functionality from the block to the
Scheduler. Since then the tag handling was changed considerably, so that
in the meantime this line can be safely removed.

Thanks to @mattkretz for spotting this and starting the initial
investigation.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>