Refactor gr::registerBlock
1. A simple `registerBlock` overload for a list of complete block types
was missing. If the given type is a specialized template, then the
`type_name` is transformed into a template argument string and passed to
`addBlockType`.

2. Generalize the `registerBlock` function for a block template with
type template arguments to an arbitrary number of parameters/arguments.

3. Generalize the `registerBlockTT` function to an arbitrary number of
type lists (and corresponding number of parameters for the template
template parameter). The relatively expensive use of std::apply is
replaced by an `outer_product` over all given typelists.

4. Add a comment on how to improve the remaining `registerBlock`
overloads.

5. Add meta::outer_product (lifted from mattkretz/virtest).

Signed-off-by: Matthias Kretz <m.kretz@gsi.de>