qa_Messages: increase timeouts and allow retry
The qa_Messages::"message passing via scheduler" test reproducibly fails
on the (relatively resource constrained) CI runners for the scheduler's
blocking execution strategy.
From the test output it could be observed, that the first get would
timeout, then the set would fail because it would unexpectedly find the
reply of the get command and then the second get would fail because it
receives the reply of the first get command.
Because `checkReply` failed with an expect assertion instead of
returning false, the retry parameter of the test was not working and the
test would fail.

This increases the timeout for the tests from 2 to 4 seconds, reduces
the timeout for the test that expects to timeout, to not increase
execution times unnecessarily and changes the assertion of the received
message contents to returning false.

Also I noticed the test will actually pass regardless of whether the
test was sucessful which makes diagnosing harder than needed so i added
an additional `mayFail` test parameter and added an assertion that the
test actually passed successfully.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>