Emscripten: disable fmt consteval functionality
The version of fmt we are using seems to have problems with the latest
versions of emscripten and/or the compiler/stdlib combination it is
using. By setting FMT_CONSTEVAL= the error can be prevented.
This syntax looks a bit weird, but is actually correct for setting the
define to an empty string.

Probably this is a measure that is a bit too drastic, but my assumption
is, that on the old emscripten it would just not detect consteval
support at all.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>