Ccache followup (#519)
* Install ccache in Dockerfile instead of action

Removes the need to reinstall ccache in each run of the action and
simplifies the build file.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>

* CI: fix enabling coverage target for debug build

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>

* CI: ccache is already set-up in the CMakeLists

Our cmake setup already contains logic to automatically use ccache if
available.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>

* CI: use absolue path to explicitly set ccache directory

The path containing the tilde was not uploaded correctly.
Also explicitily set the path to the ccache cache via environment
variable.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>

---------

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>