Implement handling "set graph YAML" message
We move all the existing blocks to the zombie list. They'll get
deleted when stopped.

The scheduler needs to be restarted, so its init() method runs again
to construct the job list and threads.

The message is being processed by a worker thread, therefore we can't
restart the scheduler here as that would crash. Scheduler is restarted
automatically by OD when it receives the message reply.

Also, the scheduler didn't support being restarted at all, the reset()
method implementation was missing.

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>