tests: Wrap sendMessage() with our own testing::sendAndWaitMessage()
There's lots of duplicated code that's done after sending a message,
like waiting for replies, which is spread throughout all tests. Not
only duplicated, sometimes it's wrong or flaky, and sometimes we forget
to wait, relying on racy behaviour.

Besides reducing the code in the tests and making it more fool-proof,
it also ports away from counting number of reply messages. As there
might be unrelated messages in the queue. Instead, wait for specific
messages, by comparing endpoint. More complex comparisons can be passed
as argument.

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>