Add targetGraph property to kEmplaceBlock message
Since the kEmplaceBlock message handling was moved to Scheduler,
it was not possible to add new blocks into a subgraph using
messages.

This adds a targetGraph property which, when defined,
tells the scheduler into which of its subgraphs the block
should be emplaced. If it is not defined, then the block
is added to the main graph.

Signed-off-by: Ivan Čukić <ivan.cukic@kdab.com>