BasicBlocks: Also add python dependency to header targets
Use of the headers also implies the dependency on python, otherwise
every user of the block library has to specify the python dependency
manually or gets errors if the python library can not be loaded.

Also use the imported targets instead manually adding paths from the
findPackage populated variables and remove the unused define and global
include directories.

Signed-off-by: Alexander Krimm <A.Krimm@gsi.de>